+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 - `Innmind\Filesystem\Adapter\InMemory::new()`
 - `Innmind\Filesystem\Adapter\Filesystem::withCaseSensitivity()`, case sensitivity can be specified as the second argument of `::mount()`
+- `Innmind\Filesystem\Exception\*`
 
 ## 8.1.0 - 2025-05-09